server/requirements-server.txt
fastapi
pydantic
uvicorn
httpx
python-dotenv
pydantic-settings
openai
ngrok